Crate revive_strategy

This crate provides the Revive strategy for the Foundry EVM ExecutorStrategy.

It is designed to work with the Revive runtime, allowing for the execution of smart contracts in a Polkadot environment.

It is heavily inspired from https://github.com/matter-labs/foundry-zksync/tree/main/crates/strategy/zksync

Modules

backend 🔒

cheatcodes 🔒

executor 🔒

link

state 🔒

tracing 🔒

Enums

ReviveRuntimeMode

Runtime backend mode for pallet-revive

ReviveStartupMigration

Controls the automatic migration to pallet-revive during test execution.

Traits

PvmCheatcodeInspectorStrategyBuilder

ReviveExecutorStrategyBuilder

Create Revive strategy for [ExecutorStrategy].