revive_strategy::cheatcodes

Function migrate_contract_storage

fn migrate_contract_storage(
    data: Ecx<'_, '_, '_>,
    address: Address,
    account_h160: H160,
)

Migrates contract storage from REVM state to pallet-revive.

Merges storage from two sources:

Journaled state: most recent storage values from current execution

Database cache: storage written before startup migration, which is run as separate transaction, already committed to cache

The journaled state takes precedence - cache values are only used for slots not present in the journaled state.